NCI-Agency/anet#2926: Rename AggregationWidgets to AggregationWidget
The file is only exporting the component with the same name, as
singular.
