+++ client/src/components/AggregationWidget.js
